+++ internal/veracode-api-wrapper/update-mitigation-info/scripts/updatemitigationinfo.py
@@ -1,0 +1,108 @@
+#
+# Copyright (c) 2019, WSO2 Inc., WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#
+# WSO2 Inc. licenses this file to you under the Apache License,
+# Version 2.0 (the "License"); you may not use this file except
+# in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+import argparse
+import csv
+import requests
+import re
+import sys
+
+GET_BUILD_API = 'https://analysiscenter.veracode.com/api/5.0/getbuildlist.do'
+UPDATE_MITIGATION_API = 'https://analysiscenter.veracode.com/api/updatemitigationinfo.do'
+BUILD_REGEX = '(build build_id=\")(\d+)'
+SUCCESS_RESPONSE_KEY = 'desc='
+PREFIX = '[VERACODE API WRAPPER] '
+ACTION_FALSEPOSITIVE = 'fp'
+ACTION_APPDESIGN = 'appdesign'
+
+print PREFIX, '- START'
+parser = argparse.ArgumentParser(description='Veracode API wrapper to update mitigation comment.')
+parser.add_argument('-a', action='store', dest='applicationID',
+                    help='Veracode Application ID')
+
+parser.add_argument('-b', action='store', dest='buildID',
+                    help='Build ID of the scan')
+
+parser.add_argument('-u', type=str, action='store', dest='username', required=True,
+                    help='Username')
+
+parser.add_argument('-cf', type=str, action='store', dest='csvFilePath', required=True,
+                    help='CSV file path of feed back report')
+
+parser.add_argument('-of', type=str, action='store', dest='outputFilePath', required=True,
+                    help='Output file path')
+
+args = parser.parse_args()
+
+password = raw_input("Enter password: ")
+
+if (args.applicationID == None):
+    print PREFIX, '- Application ID is not provided. Either Application ID or Build Id should be provided.'
+    sys.exit()
+
+# Get build id of particular application if build id is not provided.
+if (args.buildID == None):
+    print PREFIX, '- Retrieving the build id of application id : ', args.applicationID
+    payload = {'app_id': args.applicationID}
+    response = requests.post(GET_BUILD_API, params=payload, files=None, auth=(args.username, password))
+    if response.status_code == '200':
+        matchObj = re.search(BUILD_REGEX, response.text, re.M | re.I)
+        if matchObj:
+            args.buildID = matchObj.group(2)
+            print PREFIX, "- Build Id of given application : " + args.buildID
+        else:
+            print PREFIX, "- Error occurred while retrieving build ID of given application."
+            sys.exit()
+    else:
+        print PREFIX, "- Error occurred while invoking the API to retrieve build ID of given application. " \
+                      "Response code :" + response.status_code
+        sys.exit()
+
+# Update mitigation comment.
+def updateComment(flowId, comment, action):
+    updateCommentPayload = {'build_id': args.buildID, 'action': action, 'comment': comment, 'flaw_id_list': flowId}
+    print updateCommentPayload
+    response = requests.post(UPDATE_MITIGATION_API, params=updateCommentPayload, files=None,
+                             auth=(args.username, password))
+    if response.status_code == '200':
+        if SUCCESS_RESPONSE_KEY not in response.text:
+            output = "SUCCESS , Response : ", response.content
+        else:
+            output = "ERROR , Response : ", response.content
+    else:
+        output = "ERROR , Response : ", response.content
+    return output
+
+print PREFIX, "- Reading the CSV entries from analysed report"
+
+# Read CSV file and call api to update mitigation comment.
+with open(args.csvFilePath, 'r') as csvfile:
+    reader = csv.DictReader(csvfile)
+    with open(args.outputFilePath, 'a') as outputFile:
+        for row in reader:
+            flowId = dict(row).get("issue_id")
+            if (dict(row).get("WSO2_resolution") == 'False Positive'):
+                action = ACTION_FALSEPOSITIVE
+            else:
+                action = ACTION_APPDESIGN
+            comment = "Resolution: " + dict(row).get("WSO2_resolution") + "\n\n" + \
+                      "Use case: " + dict(row).get("Use_Case") + "\n\n" \
+                      "Vulnerability Influence: " + dict(row).get("Vulnerability_Influence")
+            output = updateComment(flowId, comment, action)
+            outputFile.write("Flow ID : " + flowId + " " + output)
+            file.write("\n")
