--- internal/veracode-api-wrapper/updatemitigationinfo/requirements.txt
@@ -1,3 +1,0 @@
-requests
-argparse
-regex
